updated session reducer
