+++ package.json
@@ -12,6 +12,7 @@
     "babel-jest": "20.0.3",
     "babel-loader": "7.1.2",
     "babel-preset-react-app": "^3.1.1",
+    "babel-preset-stage-0": "^6.24.1",
     "babel-runtime": "6.26.0",
     "bootstrap": "^4.1.1",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
@@ -61,7 +62,8 @@
   "scripts": {
     "start": "node scripts/start.js",
     "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
+    "test": "node scripts/test.js --env=jsdom",
+    "test:jest": "jest"
   },
   "jest": {
     "collectCoverageFrom": [
@@ -71,8 +73,7 @@
       "<rootDir>/config/polyfills.js"
     ],
     "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs}",
-      "<rootDir>/src/**/?(*.)(spec|test).{js,jsx,mjs}"
+      "<rootDir>/tests/**/?(*.)(spec|test).{js,jsx,mjs}"
     ],
     "testEnvironment": "node",
     "testURL": "http://localhost",
@@ -99,6 +100,7 @@
   },
   "babel": {
     "presets": [
+      "stage-0",
       "react-app"
     ]
   },
