+++ src/reducers/session.js
@@ -1,10 +1,30 @@
-import {ACTIONS} from '../constants/session';
+import { ACTIONS } from '../constants/session';
 
-const initialState = {
+const defaultState = {
     user: null,
-};
-const sessionReducer = (state = initialState, {type, payload}) => {
+}
+
+function setLoadingState(presState, target) {
+    return {
+        ...presState, 
+        [target]: {
+            isLoading: true,
+        }
+    };
+}
+
+function storeUser(presState, user) {
+    return {
+        ...presState,
+        user,
+    };
+}
+
+const sessionReducer = (state = defaultState, {type, payload}) => {
     switch (type) {
+        case ACTIONS.GET_USER: return setLoadingState(state, 'user');
+        case ACTIONS.STORE_USER: return storeUser(state, payload);
+        case ACTIONS.AUTH_FAILED: return storeUser(state, null);
         default: return state;
     }
 };
