+++ src/services/fontAwesome.js
@@ -1,7 +1,7 @@
 import { library } from '@fortawesome/fontawesome-svg-core';
-import { faWallet, faAddressBook, google } from '@fortawesome/free-solid-svg-icons';
+import { faWallet, faAddressBook } from '@fortawesome/free-solid-svg-icons';
 
-const icons = [faWallet, faAddressBook, google]
+const icons = [faWallet, faAddressBook]
 
 export function initialize() {
     return library.add(...icons);
